requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
joblib==1.3.2
pandas==2.1.1
scikit-learn==1.3.2
xgboost==2.0.0
openai==1.3.0
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.2
aiofiles==23.2.1
geopandas==0.14.1
shapely==2.0.1
pyproj==3.6.1
fiona==1.9.5
matplotlib==3.8.0
pyarrow==14.0.1 
